:memo: Day13优化代码,增加注释,压缩安装包的体积

+++ app/src/main/java/com/example/openeye/logic/model/CategoryDetailBean.kt
@@ -221,7 +221,7 @@ data class CategoryDetailBean(
                     ) {
                         data class Url(
                             val name: String,
-                            val size: Int,
+                            val size: Long,
                             val url: String
                         )
                     }
@@ -299,7 +299,7 @@ data class CategoryDetailBean(
             ) {
                 data class Url(
                     val name: String,
-                    val size: Int,
+                    val size: Long,
                     val url: String
                 )
             }
